Centralize code snippets in dedicated file to reduce duplication
